+++ src/SMAPI.Toolkit/Framework/Clients/CompatibilityRepo/RawDataModels/RawModUnofficialUpdate.cs
@@ -1,7 +1,7 @@
-namespace StardewModdingAPI.Toolkit.Framework.Clients.CompatibilityRepo.Internal.DataModels;
+namespace StardewModdingAPI.Toolkit.Framework.Clients.CompatibilityRepo.RawDataModels;
 
 /// <summary>As part of <see cref="RawModEntry"/>, an unofficial update which fixes compatibility with the latest Stardew Valley and SMAPI versions.</summary>
-internal class RawModUnofficialUpdate
+public class RawModUnofficialUpdate
 {
     /// <summary>The version of the unofficial update.</summary>
     public string? Version { get; set; }